requirements.txt
asgiref==3.6.0
black==22.12.0
boto3==1.26.59
botocore==1.29.59
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.2.0
Django==4.1.5
django-cors-headers==3.13.0
django-oauth-toolkit==2.2.0
django-storages==1.13.2
djangorestframework==3.14.0
drf-social-oauth2==1.2.1
ecdsa==0.18.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
jwcrypto==1.4.2
mypy-extensions==0.4.3
oauthlib==3.2.2
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
psycopg2==2.9.5
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.14
wrapt==1.14.1
